# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  /**
     * @minLength 3
     * @maxLength 32
     */
  username: string;
  email?: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginRequest {
  login: string;
  password: string;
}

export type UserResponseRole = typeof UserResponseRole[keyof typeof UserResponseRole];


export const UserResponseRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserResponseStatus = typeof UserResponseStatus[keyof typeof UserResponseStatus];


export const UserResponseStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface UserResponse {
  id: string;
  username: string;
  email?: string;
  walletAddress?: string;
  role: UserResponseRole;
  status: UserResponseStatus;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserResponse;
}

export interface UpdateWalletRequest {
  walletAddress: string;
}

export interface BalanceResponse {
  available: string;
  locked: string;
  currency: string;
}

export interface BetSelectionRequest {
  matchId: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  market: string;
  selection: string;
  odds: number;
  commenceTime?: string;
}

export interface CreateBetRequest {
  stake: number;
  /** @minItems 1 */
  selections: BetSelectionRequest[];
}

export type BetResponseSelectionsItem = { [key: string]: unknown };

export interface BetResponse {
  id: string;
  stake: string;
  totalOdds: string;
  potentialReturn: string;
  status: string;
  currency: string;
  createdAt: string;
  selections?: BetResponseSelectionsItem[];
}

export interface BetsListResponse {
  bets: BetResponse[];
}

export interface AdminUsersResponse {
  users: UserResponse[];
  total: number;
}

export interface AdminBetsResponse {
  bets: BetResponse[];
  total: number;
}

export type UpdateStatusRequestStatus = typeof UpdateStatusRequestStatus[keyof typeof UpdateStatusRequestStatus];


export const UpdateStatusRequestStatus = {
  active: 'active',
  suspended: 'suspended',
  banned: 'banned',
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}

export interface CommissionSettingItem {
  level: number;
  rate: string;
}

export interface CommissionSettingsResponse {
  settings: CommissionSettingItem[];
}

export interface UpdateCommissionSettingsRequest {
  /** @minItems 1 */
  settings: CommissionSettingItem[];
}

export interface SettlementRunResult {
  checked: number;
  settled: number;
  won: number;
  lost: number;
  void: number;
  skipped: number;
  totalPaidOut: number;
}

export interface SettlementStatsResponse {
  total: number;
  pending: number;
  won: number;
  lost: number;
  void: number;
  totalWagered: string;
  totalPaidOut: string;
  houseEdge: string;
}

export type SettleBetRequestStatus = typeof SettleBetRequestStatus[keyof typeof SettleBetRequestStatus];


export const SettleBetRequestStatus = {
  won: 'won',
  lost: 'lost',
  void: 'void',
} as const;

export interface SettleBetRequest {
  status: SettleBetRequestStatus;
}

export interface TransactionItem {
  id: string;
  userId: string;
  type: string;
  amount: string;
  currency: string;
  status: string;
  reference?: string;
  description?: string;
  createdAt: string;
}

export interface AdminTransactionsResponse {
  transactions: TransactionItem[];
  total: number;
}