# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: bets
    description: Betting operations
  - name: balance
    description: Balance operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me/wallet:
    patch:
      operationId: updateWallet
      tags: [users]
      summary: Update connected wallet address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWalletRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Wallet already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/me/balance:
    get:
      operationId: getBalance
      tags: [balance]
      summary: Get user balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"

  /users/me/bets:
    get:
      operationId: getUserBets
      tags: [bets]
      summary: Get user bet history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of bets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetsListResponse"

  /bets:
    post:
      operationId: createBet
      tags: [bets]
      summary: Place a bet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBetRequest"
      responses:
        "201":
          description: Bet placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"

  /admin/bets:
    get:
      operationId: adminGetBets
      tags: [admin]
      summary: List all bets (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bets list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBetsResponse"

  /admin/users/{id}/status:
    patch:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Update user status (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

  /admin/bets/{id}/settle:
    patch:
      operationId: adminSettleBet
      tags: [admin]
      summary: Manually settle a bet (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleBetRequest"
      responses:
        "200":
          description: Settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/transactions:
    get:
      operationId: adminGetTransactions
      tags: [admin]
      summary: List all transactions (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionsResponse"

  /admin/settlement/run:
    post:
      operationId: adminRunSettlement
      tags: [admin]
      summary: Trigger auto-settlement run (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settlement run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementRunResult"

  /admin/settlement/stats:
    get:
      operationId: adminGetSettlementStats
      tags: [admin]
      summary: Get settlement statistics (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settlement stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementStatsResponse"

  /admin/commission-settings:
    get:
      operationId: adminGetCommissionSettings
      tags: [admin]
      summary: Get commission settings (admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Commission settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSettingsResponse"
    put:
      operationId: adminUpdateCommissionSettings
      tags: [admin]
      summary: Update commission settings (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommissionSettingsRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSettingsResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required: [username, password]

    LoginRequest:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
      required: [login, password]

    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        walletAddress:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, suspended, banned]
        createdAt:
          type: string
          format: date-time
      required: [id, username, role, status, createdAt]

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required: [token, user]

    UpdateWalletRequest:
      type: object
      properties:
        walletAddress:
          type: string
      required: [walletAddress]

    BalanceResponse:
      type: object
      properties:
        available:
          type: string
        locked:
          type: string
        currency:
          type: string
      required: [available, locked, currency]

    BetSelectionRequest:
      type: object
      properties:
        matchId:
          type: string
        sport:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        market:
          type: string
        selection:
          type: string
        odds:
          type: number
        commenceTime:
          type: string
      required: [matchId, sport, homeTeam, awayTeam, market, selection, odds]

    CreateBetRequest:
      type: object
      properties:
        stake:
          type: number
        selections:
          type: array
          items:
            $ref: "#/components/schemas/BetSelectionRequest"
          minItems: 1
      required: [stake, selections]

    BetResponse:
      type: object
      properties:
        id:
          type: string
        stake:
          type: string
        totalOdds:
          type: string
        potentialReturn:
          type: string
        status:
          type: string
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
        selections:
          type: array
          items:
            type: object
      required: [id, stake, totalOdds, potentialReturn, status, currency, createdAt]

    BetsListResponse:
      type: object
      properties:
        bets:
          type: array
          items:
            $ref: "#/components/schemas/BetResponse"
      required: [bets]

    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/UserResponse"
        total:
          type: integer
      required: [users, total]

    AdminBetsResponse:
      type: object
      properties:
        bets:
          type: array
          items:
            $ref: "#/components/schemas/BetResponse"
        total:
          type: integer
      required: [bets, total]

    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
      required: [status]

    CommissionSettingItem:
      type: object
      properties:
        level:
          type: integer
        rate:
          type: string
      required: [level, rate]

    CommissionSettingsResponse:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/CommissionSettingItem"
      required: [settings]

    UpdateCommissionSettingsRequest:
      type: object
      properties:
        settings:
          type: array
          items:
            $ref: "#/components/schemas/CommissionSettingItem"
          minItems: 1
      required: [settings]

    SettlementRunResult:
      type: object
      properties:
        checked:
          type: integer
        settled:
          type: integer
        won:
          type: integer
        lost:
          type: integer
        void:
          type: integer
        skipped:
          type: integer
        totalPaidOut:
          type: number
      required: [checked, settled, won, lost, void, skipped, totalPaidOut]

    SettlementStatsResponse:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        won:
          type: integer
        lost:
          type: integer
        void:
          type: integer
        totalWagered:
          type: string
        totalPaidOut:
          type: string
        houseEdge:
          type: string
      required: [total, pending, won, lost, void, totalWagered, totalPaidOut, houseEdge]

    SettleBetRequest:
      type: object
      properties:
        status:
          type: string
          enum: [won, lost, void]
      required: [status]

    TransactionItem:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
        amount:
          type: string
        currency:
          type: string
        status:
          type: string
        reference:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, amount, currency, status, createdAt]

    AdminTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionItem"
        total:
          type: integer
      required: [transactions, total]