# lib/api-zod/src/generated/types/settlementRunResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SettlementRunResult {
  checked: number;
  settled: number;
  won: number;
  lost: number;
  void: number;
  skipped: number;
  totalPaidOut: number;
}