# lib/api-zod/src/generated/types/settlementStatsResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SettlementStatsResponse {
  total: number;
  pending: number;
  won: number;
  lost: number;
  void: number;
  totalWagered: string;
  totalPaidOut: string;
  houseEdge: string;
}